requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-model-utils==4.3.1
django-recaptcha==3.0.0
idna==3.4
psycopg2==2.9.5
requests==2.28.2
sqlparse==0.4.3
stripe==5.1.0
tzdata==2022.7
urllib3==1.26.14
